Improved terminology count algorithm with case insensitive regex

+++ patHistory/src/main/java/com/mediscreen/abernathyapp/patHistory/services/PatHistoryService.java
@@ -1,7 +1,5 @@
 package com.mediscreen.abernathyapp.patHistory.services;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.stereotype.Service;
 
 import java.util.Set;
@@ -9,6 +7,6 @@ import java.util.Set;
 @Service
 public interface PatHistoryService {
 
-    int terminologySearch(String patientId, Set<String> terminology);
+    long terminologySearch(String patientId, Set<String> terminology);
 
 }
